+++ src/com/jpexs/decompiler/flash/gui/generictageditors/FloatEditor.java
@@ -16,6 +16,7 @@
  */
 package com.jpexs.decompiler.flash.gui.generictageditors;
 
+import com.jpexs.decompiler.flash.easygui.properties.PropertyEditor;
 import com.jpexs.decompiler.flash.ecma.EcmaScript;
 import com.jpexs.helpers.Helper;
 import com.jpexs.helpers.ReflectionTools;
@@ -41,7 +42,9 @@ public class FloatEditor extends JTextField implements GenericTagEditor {
     private final Class<?> type;
 
     private String fieldName;
-
+    
+    private ValueNormalizer normalizer;
+    
     @Override
     public boolean getScrollableTracksViewportWidth() {
         return true;
@@ -77,7 +80,10 @@ public class FloatEditor extends JTextField implements GenericTagEditor {
     public void reset() {
         try {
             Object val = ReflectionTools.getValue(obj, field, index);
-            setText(val == null ? "" : val.toString());
+            if (normalizer != null) {
+                val = normalizer.toViewValue(val);
+            }
+            setText(val == null ? "" : EcmaScript.toString(val));            
         } catch (IllegalArgumentException | IllegalAccessException ex) {
             // ignore
         }
@@ -86,7 +92,11 @@ public class FloatEditor extends JTextField implements GenericTagEditor {
     @Override
     public boolean save() {
         try {
-            String oldValue = (String) EcmaScript.toString(ReflectionTools.getValue(obj, field, index));
+            Object oldFieldValue = ReflectionTools.getValue(obj, field, index);
+            if (normalizer != null) {
+                oldFieldValue = normalizer.toViewValue(oldFieldValue);
+            }
+            String oldValue = (String) EcmaScript.toString(oldFieldValue);
             String newValue = getText();
             if (Objects.equals(oldValue, newValue)) {
                 return false;
@@ -97,6 +107,9 @@ public class FloatEditor extends JTextField implements GenericTagEditor {
             } else {
                 val = Float.valueOf(getText());
             }
+            if (normalizer != null) {
+                val = normalizer.toFieldValue(val);
+            }
             ReflectionTools.setValue(obj, field, index, val);
         } catch (IllegalArgumentException | IllegalAccessException ex) {
             // ignore
@@ -106,7 +119,7 @@ public class FloatEditor extends JTextField implements GenericTagEditor {
 
     @Override
     public void addChangeListener(final ChangeListener l) {
-        final GenericTagEditor t = this;
+        final PropertyEditor t = this;
         addFocusListener(new FocusAdapter() {
 
             @Override
@@ -150,4 +163,10 @@ public class FloatEditor extends JTextField implements GenericTagEditor {
     public Object getObject() {
         return obj;
     }
+    
+    @Override
+    public void setValueNormalizer(ValueNormalizer normalizer) {
+        this.normalizer = normalizer;
+        reset();
+    }  
 }